// Modulo_03/CineTrack/src/script/ts/types.ts
// Enums
export type Genero =
  | "Ação"
  | "Comédia"
  | "Drama"
  | "Ficção Cientifica"
  | "Terror"
  | "Romance"
  | "Animação"
  | "Aventura";

export type Status =
  | "Quero ver"
  | "Assistindo"
  | "Concluído";

export type TipoTitulo = "Filme" | "Série";

export type Plataforma =
  | "Netflix"
  | "Prime Video"
  | "Disney+"
  | "HBO Max";